I do small steps since it is basically impossible to know what goes wrong once it does...

+++ src/freert/planar/ShapePolygon.java
@@ -400,13 +400,13 @@ public final class ShapePolygon extends ShapePolyline
       PlaLineInt curr_line = new PlaLineInt(corners[0], corners[1]);
       PlaDirection first_direction = first_line.direction();
       PlaDirection curr_direction = curr_line.direction();
-      double last_det = first_direction.determinant(curr_direction);
+      long last_det = first_direction.determinant(curr_direction);
 
       for (int ind2 = 2; ind2 < corners.length; ++ind2)
          {
          curr_line = new PlaLineInt(curr_line.point_b, corners[ind2]);
          curr_direction = curr_line.direction();
-         double curr_det = first_direction.determinant(curr_direction);
+         long curr_det = first_direction.determinant(curr_direction);
          if (last_det <= 0 && curr_det > 0)
             return false;
          last_det = curr_det;
